teal module for the AE by subgroups

Display the AE by subgroups plot as a teal module

Usage

tm_g_ae_sub(
  label,
  dataname,
  parentname = "ADSL",
  arm_var = teal.picks::variables(dplyr::starts_with("ACTARM")),
  group_var = teal.picks::variables(is.factor),
  plot_height = c(600L, 200L, 2000L),
  plot_width = NULL,
  fontsize = c(5, 3, 7),
  transformators = list(),
  decorators = list()
)

Arguments

label

(character(1)) Label shown in the navigation item for the module or module group. For modules() defaults to "root". See Details.

dataname

(character(1))
analysis data used in the teal module, needs to be available in the list passed to the data argument of teal::init().

parentname

(character(1))
analysis data used for several variables in the teal module, needs to be available in the list passed to the data argument of teal::init(). The default is "ADSL"

arm_var

Either a (teal.picks::variables()) object or a (teal.transform::choices_selected()) object.
choices_selected is being deprecated as an argument type and will be removed in the future. Object with all available choices and the pre-selected option for variable names that can be used as arm_var. Column arm_var in the dataname has to be a factor.

group_var

Either a (teal.picks::variables()) object or a (teal.transform::choices_selected()) object. choices_selected() is being deprecated as an argument type and will be removed in the future. Object with subgroup variables.

plot_height

(numeric(3))
vector to indicate default value, minimum and maximum values.

plot_width

(numeric(3))
vector to indicate default value, minimum and maximum values.

fontsize

(numeric(1) or numeric(3))
Defines initial possible range of font-size. fontsize is set for teal.widgets::optionalSliderInputValMinMax() which controls font-size in the output plot.

transformators

(list of teal_transform_module) that will be applied to transform module's data input. To learn more check vignette("transform-input-data", package = "teal").

decorators

(named list of teal_transform_module) optional, decorators for the module plot output.

Value

the teal::module() object.

Decorating Module

This module generates the following objects, which can be modified in place using decorators:

plot (grob)

A Decorator is applied to the specific output using a named list of teal_transform_module objects. The name of this list corresponds to the name of the output to which the decorator is applied. See code snippet below:

tm_g_ae_sub(
   ..., # arguments for module
   decorators = list(
     plot = teal_transform_module(...), # applied to the `plot` output
   )
)

For additional details and examples of decorators, refer to the vignette vignette("decorate-module-output", package = "teal.modules.general").

To learn more please refer to the vignette vignette("transform-module-output", package = "teal") or the teal::teal_transform_module() documentation.

Reporting

This module returns an object of class teal_module, that contains a server function. Since the server function returns a teal_report object, this makes this module reportable, which means that the reporting functionality will be turned on automatically by the teal framework.

For more information on reporting in teal, see the vignettes:

vignette("reportable-shiny-application", package = "teal.reporter")

vignette("adding-support-for-reporting-to-custom-modules", package = "teal")

Author

Liming Li (Lil128) liming.li@roche.com

Molly He (hey59) hey59@gene.com

Examples

# Example using stream (ADaM) dataset
data <- within(teal_data(), {
  ADSL <- teal.data::rADSL
  ADAE <- teal.data::rADAE
})

join_keys(data) <- default_cdisc_join_keys[names(data)]

app <- init(
  data = data,
  modules = modules(
    tm_g_ae_sub(
      label = "AE by Subgroup",
      dataname = "ADAE",
      parentname = "ADSL",
      arm_var = variables(
        choices = c("ACTARM", "ACTARMCD"),
        selected = "ACTARMCD"
      ),
      group_var = variables(
        choices = c("SEX", "REGION1", "RACE"),
        selected = c("SEX", "REGION1", "RACE"),
        multiple = TRUE
      )
    )
  )
)
#> Initializing tm_g_ae_sub
if (interactive()) {
  shinyApp(app$ui, app$server)
}