Adverse Events Table Teal Module

This shiny module displays the adverse events table (AET02).

tm_t_ae(label, dataname, filter_var = NULL, arm_var, class_var, term_var,
  total_col = TRUE, pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init.

filter_var
choices_selected variable name of data filter, please see details regarding expected values, default is NULL

arm_var
choices_selected variable name in analysis data that is used as col_by argument for the respective tern or osprey function.

class_var
choices_selected class variables selected for display

term_var
choices_selected term variables selected for display

total_col
argument for appearance of "All Patients" column (default is TRUE)

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized, default is NULL

Value

an module object

Details

filter_var option is designed to work in conjunction with filtering function provided by teal (encoding panel on the right hand side of the shiny app). It can be used as quick access to predefined subsets of the domain datasets (not subject-level dataset) to be used for analysis, denoted by an value of "Y". Each variable within the filter_var_choices is expected to contain values of either "Y" or "N". If multiple variables are selected as filter_var, only observations with "Y" value in each and every selected variables will be used for subsequent analysis. Flag variables (from ADaM datasets) can be used directly as filter.

Examples

#Example using stream (ADaM) dataset
library(dplyr)

ASL <- rADSL %>% mutate(USUBJID = SUBJID)
AAE <- rADAE %>% mutate(flag1 = ifelse(SEX == "F", "Y", "N"))  %>% mutate(USUBJID = SUBJID)

app <- init(
  data = cdisc_data(
    ASL = ASL,
    AAE = AAE,
    code = paste0(c(
      "ASL <- rADSL %>% mutate(USUBJID = SUBJID)",
      'AAE <- rADAE %>% mutate(flag1 = ifelse(SEX == "F", "Y", "N")) %>% mutate(USUBJID = SUBJID)'),
      collapse = ";"
    ),
    check = FALSE
  ),
  modules = root_modules(
    tm_t_ae(
      label = "Adverse Events Table",
      dataname = "AAE",
      filter_var = choices_selected(
        choices = c("DTHFL", "flag1"),
        selected = NULL
      ),
      arm_var = choices_selected(
        choices = c("ARM", "ARMCD"),
        selected = "ARM"
      ),
      class_var = choices_selected(
        choices = c("AEBODSYS", "AEHLTCD"),
        selected = "AEBODSYS"
      ),
      term_var = choices_selected(
        choices = c("AEDECOD", "AETERM"),
        selected = "AEDECOD"
      ),
      total_col = TRUE
    )
  )
)

#> |apply filter for ASL
#> |apply filter for AAE
#> |** UI id for modules is teal_modules_root
#> |UI id for module is teal_modules_root_Adverse_Events_Table

if (FALSE) {
shinyApp(app$ui, app$server)
}