Teal module for the AE overview

Display the AE overview plot as a shiny module

tm_g_ae_oview(
  label,
  dataname,
  arm_var,
  add_flag = NULL,
  fontsize = c(5, 3, 7),
  plot_height = c(600L, 200L, 2000L),
  plot_width = NULL
)

Arguments

label

(character)
menu item label of the module in the teal app

dataname

(character) analysis data used in the teal module, needs to be available in the list passed to the data argument of init.

arm_var

choices_selected object with all available choices and the pre-selected option for variable names that can be used as arm_var.

add_flag

choices_selected, a string or a vector of characters including variable name(s) for additional flags, default is NULL (i.e. no additional flags will be added)

fontsize

a numeric vector with 3 values, selected font size and font size range, default is c(5, 3, 7)

plot_height

`numeric` vector to indicate default value, minimum and maximum values.

plot_width

`numeric` vector to indicate default value, minimum and maximum values.

Value

a module object

Examples

library(scda)
#> 

ADSL <- synthetic_cdisc_data("latest")$adsl
ADAE <- synthetic_cdisc_data("latest")$adae

# Add additional dummy causality flags.
ADAE <- ADAE %>%
  mutate(AEREL1 = (AEREL == "Y" & ACTARM == "A: Drug X")) %>%
  mutate(AEREL2 = (AEREL == "Y" & ACTARM == "B: Placebo")) %>%
  rtables::var_relabel(
    AEREL1 = "AE related to A: Drug X",
    AEREL2 = "AE related to B: Placebo"
  )

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"),
    cdisc_dataset("ADAE", ADAE,
      code = "ADAE <- synthetic_cdisc_data(\"latest\")$adae
              # Add additional dummy causality flags.
              ADAE <- ADAE %>%
                mutate(AEREL1 = (AEREL == 'Y' & ACTARM == 'A: Drug X')) %>%
                mutate(AEREL2 = (AEREL == 'Y' & ACTARM == 'B: Placebo')) %>%
                rtables::var_relabel(
                  AEREL1 = 'AE related to A: Drug X',
                  AEREL2 = 'AE related to B: Placebo'
                )"
    ),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_ae_oview(
      label = "AE Overview",
      dataname = "ADAE",
      arm_var = choices_selected(
        selected = "ACTARM",
        choices = c("ACTARM", "ACTARMCD")
      ),
      add_flag = choices_selected(
        choices = variable_choices(ADAE, c("AEREL1", "AEREL2")),
        selected = NULL
      ),
      plot_height = c(600, 200, 2000)
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}