Butterfly plot Teal Module

Display butterfly plot as a shiny module

tm_g_butterfly(label, dataname, filter_var = NULL,
  filter_var_choices = NULL, right_var, right_var_choices = right_var,
  left_var, left_var_choices = left_var, category_var,
  category_var_choices = category_var, color_by_var,
  color_by_var_choices = color_by_var, count_by_var,
  count_by_var_choices = count_by_var, facet_var = NULL,
  facet_var_choices = facet_var, sort_by_var = "count",
  sort_by_var_choices = c("count", "alphabetical"), legend_on = TRUE,
  plot_height, pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init.

filter_var
variable name of data filter, please see details regarding expected values, default is NULL

filter_var_choices
vector with filter_var choices, default is NULL

right_var
dichotomization variable for right side

right_var_choices
vector with dichotomization choices

left_var
dichotomization variable for left side

left_var_choices
vector with dichotomization choices

category_var
category (y axis) variable

category_var_choices
vector of category choices

color_by_var
variable defines color blocks within each bar

color_by_var_choices
vector of choices for color_block_by

count_by_var
variable defines how x axis is calculated

count_by_var_choices
vector of choices for count_by_var

facet_var
variable for row facets

facet_var_choices
vector with facet_var choices

sort_by_var
argument for order of class and term elements in table, default here is "count"

sort_by_var_choices
vector with sort_by_var choices

legend_on
boolean value for whether legend is displayed

plot_height
range of plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized, default is NULL

Value

an module object

Details

filter_var option is designed to work in conjuction with filtering function provided by teal (encoding panel on the right hand side of the shiny app). It can be used as quick access to pre-defined subsets of the domain datasets (not subject-level dataset) to be used for analysis, denoted by an value of "Y". Each variable within the filter_var_choices is expected to contain values of either "Y" or "N". If multiple variables are selected as filter_var, only observations with "Y" value in each and every selected variables will be used for subsequent analysis. Flag variables (from ADaM datasets) can be used directly as filter.

Examples

# NOT RUN {
#Example butterfly plot
library(dplyr)

data("rADSL")
data("rADAE")
ASL <- rADSL %>% mutate(DOSE = paste(sample(1:3, nrow(rADSL), replace = T), "UG"))
AAE <- rADAE
AAE <- AAE %>% mutate(flag1 = ifelse(AETOXGR == 1, 1, 0)) %>%
               mutate(flag2 = ifelse(AETOXGR == 2, 1, 0)) %>%
               mutate(flag3 = ifelse(AETOXGR == 3, 1, 0))
AAE <- AAE %>% mutate(flag1_filt = rep("Y", nrow(AAE)))

x <- teal::init(
  data = list(ASL = ASL, AAE = AAE),
  modules = root_modules(
    tm_g_butterfly(
       label = "Butterfly Plot",
       dataname = "AAE",
       right_var = "SEX",
       right_var_choices = c("DOSE","SEX", "ARM", "RACE", "flag1", "flag2", "flag3"),
       left_var = "RACE",
       left_var_choices = c("DOSE", "SEX", "ARM", "RACE", "flag1", "flag2", "flag3"),
       category_var = "AEBODSYS",
       category_var_choices = c("AEDECOD", "AEBODSYS"),
       color_by_var = "AETOXGR",
       color_by_var_choices = c("AETOXGR", "None"),
       count_by_var = "# of patients",
       count_by_var_choices = c("# of patients", "# of AEs"),
       facet_var = NULL,
       facet_var_choices = c("RACE", "SEX", "ARM"),
       sort_by_var = "count",
       sort_by_var_choices = c("count", "alphabetical"),
       legend_on = TRUE,
       plot_height = c(600, 200, 2000)
   )
  )
)

shinyApp(x$ui, x$server)

# }