Spiderplot Teal Module

Display spiderplot as a shiny module

tm_g_spiderplot(label, dataname, paramcd, paramcd_choices = paramcd,
  x_var, x_var_choices = x_var, y_var, y_var_choices = y_var,
  marker_var, marker_var_choices = marker_var, line_colorby_var,
  line_colorby_var_choices = line_colorby_var_choices,
  vref_line = NULL, href_line = NULL, anno_txt_var,
  legend_on = FALSE, xfacet_var = NULL,
  xfacet_var_choices = xfacet_var, yfacet_var = NULL,
  yfacet_var_choices = y_facet_var, plot_height, pre_output = NULL,
  post_output = NULL, code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

paramcd
single selected endpoint filtered with PARAMCD variable

paramcd_choices
vector with paramcd choices

x_var
x-axis variables

x_var_choices
vector with x_var choices

y_var
y-axis variables

y_var_choices
vector with y_var choices

marker_var
variable dictates marker symbol

marker_var_choices
vector with marker_var choices

line_colorby_var
variable dictates line color

line_colorby_var_choices
vector with line_colorby_var choices

vref_line
vertical reference lines

href_line
horizontal reference lines

anno_txt_var
annotation text

legend_on
boolean value for whether legend is displayed

xfacet_var
variable for x facets

xfacet_var_choices
vector with xfacet_var choices

yfacet_var
variable for y facets

yfacet_var_choices
vector with yfacet_var choices

plot_height
range of plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized, default is NULL

Value

an module object

Examples

# NOT RUN {
#Example spiderplot
library(dplyr)

data("rADSL")
data("rADTR")

ASL <- rADSL
ATR <- rADTR
#dat <- left_join(ATR, ASL) %>% filter(PARAMCD == "SUMTGLES") %>% as.data.frame()

# test character vector as x axis labels
#ATR$TUDY <- as.character(ATR$TUDY)

x <- teal::init(
  data = list(ASL = ASL, ATR = ATR),
  modules = root_modules(
    tm_g_spiderplot(
       label = "Spiderplot",
       dataname = "ATR",
       paramcd = "SLDINV",
       paramcd_choices = "SLDINV",
       x_var = "ADY",
       x_var_choices = "ADY",
       y_var = "PCHG",
       y_var_choices = c("PCHG", "CHG", "AVAL"),
       marker_var = "SEX",
       marker_var_choices = c("SEX", "RACE", "USUBJID"),
       line_colorby_var = "SEX",
       line_colorby_var_choices = c("SEX","USUBJID", "RACE"),
       vref_line = "10, 37",
       href_line = "-20, 0",
       anno_txt_var = TRUE,
       legend_on = FALSE,
       xfacet_var = "SEX",
       xfacet_var_choices = c("SEX", "ARM"),
       yfacet_var = "ARM",
       yfacet_var_choices = c("ARM", "SEX"),
       plot_height = c(600, 200, 2000)
   )
  )
)

shinyApp(x$ui, x$server)
# }