Spider plot Teal Module

Display spider plot as a shiny module

tm_g_spiderplot(
  label,
  dataname,
  paramcd,
  x_var,
  y_var,
  marker_var,
  line_colorby_var,
  xfacet_var = NULL,
  yfacet_var = NULL,
  vref_line = NULL,
  href_line = NULL,
  anno_txt_var = TRUE,
  legend_on = FALSE,
  plot_height = c(600, 200, 2000),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
menu item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

paramcd
single selected endpoint filtered with PARAMCD variable

x_var
x-axis variables

y_var
y-axis variables

marker_var
variable dictates marker symbol

line_colorby_var
variable dictates line color

xfacet_var
variable for x facets

yfacet_var
variable for y facets

vref_line
vertical reference lines

href_line
horizontal reference lines

anno_txt_var
annotation text

legend_on
boolean value for whether legend is displayed

plot_height
range of plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Value

an module object

Examples

#Example using stream (ADaM) dataset
library(dplyr)
library(random.cdisc.data)

ADSL <- rADSL
ADTR <- rADTR

app <- teal::init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADTR",  ADTR,
                  keys = keys(primary = c("STUDYID", "USUBJID", "PARAMCD", "AVISIT"),
                              foreign = c("STUDYID", "USUBJID"),
                              parent = "ADSL")),
    code = 'ADSL <- rADSL; ADTR <- rADTR'),
  modules = root_modules(
    tm_g_spiderplot(
      label = "Spider plot",
      dataname = "ADTR",
      paramcd = choices_selected(choices = "SLDINV", selected = "SLDINV"),
      x_var = choices_selected(choices = "ADY", selected = "ADY"),
      y_var = choices_selected(choices = c("PCHG", "CHG", "AVAL"), selected = "PCHG"),
      marker_var = choices_selected(choices = c("SEX", "RACE", "USUBJID"), selected = "SEX"),
      line_colorby_var = choices_selected(choices = c("SEX","USUBJID", "RACE"), selected = "SEX"),
      xfacet_var = choices_selected(choices = c("SEX", "ARM"), selected = "SEX"),
      yfacet_var = choices_selected(choices = c("SEX", "ARM"), selected = "ARM"),
      vref_line = "10, 37",
      href_line = "-20, 0",
      anno_txt_var = TRUE,
      legend_on = FALSE,
      plot_height = c(600, 200, 2000)
    )
  )
)

if (FALSE) {
shinyApp(app$ui, app$server)
}