Teal Module for Swimlane Plot

This is teal module that generates a swimlane plot (bar plot with markers) for ADaM data

tm_g_swimlane(label, dataname, bar_var, bar_var_choices = bar_var,
  bar_color_var = NULL, bar_color_var_choices = bar_color_var,
  sort_var = NULL, sort_var_choices = sort_var,
  marker_pos_var = NULL, marker_pos_var_choices = marker_pos_var,
  marker_shape_var = NULL, marker_shape_var_choices = marker_shape_var,
  marker_shape_opt = NULL, marker_color_var = NULL,
  marker_color_var_choices = marker_color_var, marker_color_opt = NULL,
  anno_txt_var = NULL, anno_txt_var_choices = anno_txt_var,
  vref_line = NULL, plot_height = c(1200, 400, 5000),
  pre_output = NULL, post_output = NULL, code_data_processing = NULL)

Arguments

label
item label of the module in the teal app

dataname
analysis data used for plotting, needs to be available in the list passed to the data argument of init. If no markers are to be plotted in the module, "ASL" should be the input. If markers are to be plotted, data name for the marker data should be the input

bar_var
subject-level numeric variable from dataset to plot as the bar length

bar_var_choices
vector with variable names that can be used as bar_var

bar_color_var
color by variable (subject-level)

bar_color_var_choices
vector with variable names that can be used as bar_color_var

sort_var
sort by variable (subject-level)

sort_var_choices
vector with variable names that can be used as sort_var

marker_pos_var
numeric variable for marker position from marker data (Note: make sure that marker position has the same relative start day as bar length variable bar_var)

marker_pos_var_choices
vector with variable names that can be used as marker_pos_var

marker_shape_var
marker shape variable from marker data

marker_shape_var_choices
vector with variable names that can be used as marker_shape_var

marker_shape_opt
aesthetic values to map shape values (named vector to map shape values to each name). If not NULL, please make sure this contains all posible values for marker_shape_var values, otherwise shape will be assigned by ggplot default

marker_color_var
marker color variable from marker data

marker_color_var_choices
vector with variable names that can be used as marker_color_var

marker_color_opt
aesthetic values to map color values (named vector to map color values to each name). If not NULL, please make sure this contains all posible values for marker_color_var values, otherwise color will be assigned by ggplot default

anno_txt_var
character vector with subject-level variable names that are selected as annotation

anno_txt_var_choices
vector with variable names that can be selected as anno_txt_var

vref_line
vertical reference lines

plot_height
plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Value

a module object

Examples

# NOT RUN {
library(dplyr)

data("rADSL")
data("rADRS")

ASL <- rADSL
ARS <- rADRS

ARS <- ARS %>% filter(PARAMCD == "LSTASDI" & DCSREAS == "Death") %>%
  mutate(AVALC = DCSREAS,
         ADY = EOSDY) %>%
  rbind (ARS %>% filter(PARAMCD == "OVRINV" & AVALC != "NE")) %>%
  arrange(USUBJID)

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_g_swimlane(
       label = "Swimlane Plot",
       dataname = 'ARS',
       bar_var = "TRTDURD",
       bar_var_choices = c("TRTDURD", "EOSDY"),
       bar_color_var = "EOSSTT",
       bar_color_var_choices = c("EOSSTT", "ARM", "ARMCD", "ACTARM", "ACTARMCD", "AGEGR1", "SEX"),
       sort_var = "ACTARMCD",
       sort_var_choices = c("USUBJID", "SITEID", "ACTARMCD", "TRTDURD"),
       marker_pos_var = "ADY",
       marker_pos_var_choices = c("None", "ADY"),
       marker_shape_var = "AVALC",
       marker_shape_var_choices = c("None", "AVALC", "AVISIT"),
       marker_shape_opt = c("CR" = 16, "PR" = 17, "SD" = 18, "PD" = 15, "Death" = 8),
       marker_color_var = "AVALC",
       marker_color_var_choices = c("None", "AVALC", "AVISIT"),
       marker_color_opt = c("CR" = "green", "PR" = "blue", "SD" = "goldenrod", "PD" = "red", "Death" = "black"),
       vref_line = "30, 60",
       anno_txt_var = c("ACTARM", "SEX"),
       anno_txt_var_choices = c("ARM", "ARMCD", "ACTARM", "ACTARMCD", "AGEGR1", "SEX", "RACE","COUNTRY","DCSREAS", "DCSREASP")
   )
  )
)

shinyApp(x$ui, x$server)
# }