Teal Module for Swimlane Plot

This is teal module that generates a swimlane plot for ADaM data

tm_g_swimlane(label, dataname, bar_var, bar_var_choices = bar_var,
  bar_color_var = NULL, bar_color_var_choices = bar_color_var,
  sort_var = NULL, sort_var_choices = sort_var, marker_shape_var = NULL,
  marker_shape_var_choices = marker_shape_var, marker_shape_opt = NULL,
  marker_color_var = NULL, marker_color_var_choices = marker_color_var,
  marker_color_opt = NULL, anno_txt_var = NULL,
  anno_txt_var_choices = anno_txt_var, vref_line = NULL,
  plot_height = c(1200, 400, 5000), pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)

Arguments

label
item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init.

bar_var
numeric variable from dataset to plot as the bar length

bar_var_choices
vector with variable names that can be used as bar_var

bar_color_var
color by variable

bar_color_var_choices
vector with variable names that can be used as bar_color_var

sort_var
sort by variable

sort_var_choices
vector with variable names that can be used as sort_var

marker_shape_var
marker shape variable

marker_shape_var_choices
vector with variable names that can be used as marker_shape_var

marker_shape_opt
aesthetic values to map shape values (named vector to map shape values to each name)

marker_color_var
marker color variable

marker_color_var_choices
vector with variable names that can be used as marker_color_var

marker_color_opt
aesthetic values to map color values (named vector to map color values to each name)

anno_txt_var
character vector with variable names that are selected as annotation by default

anno_txt_var_choices
vector with variable names that can be selected as anno_txt_var

vref_line
vertical reference line

plot_height
plot height

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Value

an module object

Details

This modules expects that the analysis data has the following variables

ADY	analysis study day

ADY

Examples

# NOT RUN {
library(dplyr)

atx <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/atx.sas7bdat")
asl <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/asl.sas7bdat")
xars <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/xars.sas7bdat")
# phase 1A data
# pre-process ASl and ARS to fit in ANL
ATX <- atx %>% filter(grepl("1a", APERIDC2))
ASL <- ATX %>% select(USUBJID) %>%
left_join(asl, by = "USUBJID") %>%
filter(SAFFL == "Y")

anno_txt_var <- c("ARMCD", "SEX", "CADX")
anno_txt <- ASL[, anno_txt_var]

ARS <- ASL %>% select(USUBJID) %>%
left_join(xars %>% filter(grepl("1a", APERIDC2) & PARAMCD == "OVRINV"), "USUBJID")

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_g_swimlane(
       label = "Swimlane Plot",
       dataname = 'ARS',
       bar_var = "TRTDUR",
       bar_var_choices = c("TRTDUR", "AGE"),
       bar_color_var = "None",
       bar_color_var_choices = c("ARM", "ARMCD"),
       sort_var = "ARM",
       sort_var_choices = c("ARM", "TRTDUR"),
       marker_shape_var = "None",
       marker_shape_var_choices = c("None", "AVALC", "AVISIT"),
       marker_shape_opt = c("CR" = 16, "PR" = 17, "SD" = 18, "PD" = 15,
                            "DEATH" = 8, "LOST TO FOLLOW-UP" = 10, "WITHDRAWAL BY SUBJECT" = 14),
       marker_color_var = "None",
       marker_color_var_choices = c("None", "AVALC", "AVISIT"),
       marker_color_opt = c("CR" = "green", "PR" = "blue", "SD" = "yellow", "PD" = "red",
                            "DEATH" = "black", "LOST TO FOLLOW-UP" = "purple", "WITHDRAWAL BY SUBJECT" = "darkred"),
       vref_line = NULL,
       anno_txt_var = c("SEX", "COUNTRY"),
       anno_txt_var_choices = c("SEX", "RACE", "COUNTRY", "CADX")

   )
  )
)

shinyApp(x$ui, x$server)
# }